+++ CHANGELOG.md
@@ -3,15 +3,15 @@
 ## unreleased
 * BraintreeVisaCheckout
     * Add Visa Checkout to the SDK as a payment option
-    * Add `BTVisaCheckoutAddress`, `BTVisaCheckoutUserData`, `BTVisaCheckoutCardNonce`
-    * Add `BTVisaCheckoutClient`
-    * Add `BTVisaCheckoutAddress`
+    * Add `BTVisaCheckoutAddress`, `BTVisaCheckoutUserData`, `BTVisaCheckoutCardNonce`, `BTVisaCheckoutClient`
 * BraintreePayPal
   * Bug fix: `BTPayPalRequest.userPhoneNumber` could be passed as an empty string resulting in an error
   * Bug fix: Ensure that `ASWebAuthenticationSession.start` is only called with one URL at a time by blocking additional start calls until the current one completes.
   * Bug fix: Pass `unitTaxAmount` as expected in `BTPayPalLineItem`
   * Make `BTPayPalLineItem` parameters settable
   * Bug fix: only attempt to call `UIApplication.shared.open` on one URL at a time blocking other `open` calls until the current one is finished
+  * Bug fix: Ensure that the `browserLoginAlertCanceled` event is not fired off when the `ASWebAutheniticationSession` popup is disabled.
+  * Add `paypal:tokenize:browser-presentation:started` event for when the `ASWebAuthenticationSession.start` is launched.
 
 ## 6.34.0 (2025-06-18)
 * BraintreePayPal